sflock integration, refactoring and code cleanup
All attachments from the emails are now sent through sflock.
Thanks to that archive files should now be properly handled.
Other goodies include not having to do our own hashing of the files.
Code is now speparated into two functions, one to parse the email and a second one to upload to cuckoo.